requirements.txt
Flask
Jinja2
docopt
flask-sqlalchemy
flask-migrate
flask-login
python-dotenv
sqlalchemy-json